split telegram types into separate files
