+++ internal/telegram/callback_query.go
@@ -1,0 +1,8 @@
+package telegram
+
+type CallbackQuery struct {
+	ID      string   `json:"id"`
+	From    *User    `json:"from"`
+	Message *Message `json:"message,omitempty"`
+	Data    string   `json:"data,omitempty"`
+}
