+++ internal/telegram/chat.go
@@ -1,0 +1,5 @@
+package telegram
+
+type Chat struct {
+	Id int `json:"id"`
+}
