+++ internal/telegram/update.go
@@ -1,0 +1,7 @@
+package telegram
+
+type Update struct {
+	UpdateId      int            `json:"update_id"`
+	Message       Message        `json:"message"`
+	CallbackQuery *CallbackQuery `json:"callback_query,omitempty"`
+}
