+++ internal/telegram/user.go
@@ -1,0 +1,6 @@
+package telegram
+
+type User struct {
+	ID       int64  `json:"id"`
+	Username string `json:"username"`
+}
